DONE -CARGANDO VALORES DE MEDICION PARCELA DE INVENTARIO Y CAMBIANDO TIPO DE DATO EN ECUACION DE VOLUMENES FIX - DATETIME PICKER CON MES ACTUAL

+++ db/migrate/20150716130016_create_ecuacion_volumenes.rb
@@ -1,8 +1,8 @@
 class CreateEcuacionVolumenes < ActiveRecord::Migration
   def change
     create_table :ecuacion_volumenes do |t|
-      t.integer :dap_inicial
-      t.integer :dap_final
+      t.decimal :dap_inicial
+      t.decimal :dap_final
       t.string :ecuacion_volumen
       t.references :tipo_bosque, index: true
       t.timestamps
